上传联调时的代码修改

+++ model/src/main/java/org/software/model/content/vo/TagVO.java
@@ -7,7 +7,7 @@ import java.util.Date;
 
 @Data
 public class TagVO {
-    private Integer tagId;
+    private Long tagId;
     private String tagName;
     private Integer isActive;
     private Date createdAt;
